requirements_llma.txt
pinecone-client langchain
pypdf
sentence-transformers
replicate

